+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "element-plus-vite-starter",
   "private": true,
-  "version": "0.1.0",
+  "version": "2.4.0",
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc --noEmit && vite build",
@@ -23,7 +23,6 @@
     "clipboard": "^2.0.11",
     "cm-tarnation": "^0.4.0",
     "dayjs": "^1.11.11",
-    "element-plus": "^2.7.3",
     "fflate": "^0.7.4",
     "file-saver": "^2.0.5",
     "format-email": "^0.1.0",
@@ -39,6 +38,7 @@
     "@types/file-saver": "^2.0.7",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^17.0.45",
+    "@vicons/carbon": "^0.12.0",
     "@vitejs/plugin-vue": "^2.3.4",
     "@vue/compiler-sfc": "^3.4.27",
     "autoprefixer": "^10.4.19",
