+++ vite.config.ts
@@ -15,13 +15,6 @@ export default defineConfig({
       '~/': `${pathSrc}/`,
     },
   },
-  css: {
-    preprocessorOptions: {
-      scss: {
-        additionalData: `@use "~/styles/element/index.scss" as *;`,
-      },
-    },
-  },
   base: './',
   plugins: [
     vue(),
